Subida de archivos

+++ Clases/Get/EstadoProcesoTitulacion.cs
@@ -11,51 +11,51 @@ namespace Titulacion.Clases.Get
         public string NoControl { get; set; } = null!;
 
         [Display(Name = "Estado del paso 1")]
-        public ulong Paso1 { get; set; }
+        public int Paso1 { get; set; }
 
         [Display(Name = "Solicitud de Constancia de No Inconveniencia")]
-        public ulong Scni { get; set; }
+        public int Scni { get; set; }
 
         [Display(Name = "Constancia de No Inconveniencia")]
-        public ulong Cni { get; set; }
+        public int Cni { get; set; }
 
         [Display(Name = "Certificado de Licenciatura")]
-        public ulong Cl { get; set; }
+        public int Cl { get; set; }
 
         [Display(Name = "Constancia de Acreditación del Idioma Inglés")]
-        public ulong Caii { get; set; }
+        public int Caii { get; set; }
 
         [Display(Name = "Recibo de Pago")]
-        public ulong Rp { get; set; }
+        public int Rp { get; set; }
 
         [Display(Name = "Recibo de Pago Sellado")]
-        public ulong Rps { get; set; }
+        public int Rps { get; set; }
 
         [Display(Name = "Solicitud de Titulación")]
-        public ulong St { get; set; }
+        public int St { get; set; }
 
         [Display(Name = "Proyecto")]
-        public ulong Pro { get; set; }
+        public int Pro { get; set; }
 
         [Display(Name = "Solicitud de Liberación, Acesores, Sinodales o Comité Revisor")]
-        public ulong Sl { get; set; }
+        public int Sl { get; set; }
 
         [Display(Name = "Liberación de Proyecto")]
-        public ulong Lp { get; set; }
+        public int Lp { get; set; }
 
         [Display(Name = "Asignación de Acesores, Sinodales o Comité Revisor")]
-        public ulong Asnc { get; set; }
+        public int Asnc { get; set; }
 
         [Display(Name = "Orden de Impresión")]
-        public ulong Oi { get; set; }
+        public int Oi { get; set; }
 
         [Display(Name = "CURP")]
-        public ulong Curp { get; set; }
+        public int Curp { get; set; }
 
         [Display(Name = "RFC")]
-        public ulong Rfc { get; set; }
+        public int Rfc { get; set; }
 
         [Display(Name = "Certficado de Bachilleres")]
-        public ulong Cb { get; set; }
+        public int Cb { get; set; }
     }
 }
